Fix for a bug causing walking to be activated by overlapping CC on Vmangos servers.
The modern client sends the falling flag in every MOVE_SPLINE_DONE packet and this can trigger the movement rooted condition in the vmangos anticheat if the player is rooted at the end of the spline movement. Simply removing the flag should fix the issue and since the 1.12 doesn't seem to use the falling flag in the spline done packets, it shouldn't cause any problems.